backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
pandas==2.0.3
numpy==1.24.3
chromadb==0.4.24
sentence-transformers==2.2.2
huggingface_hub==0.14.1
openai==1.3.5
python-dotenv==1.0.0
requests==2.31.0
